id5

+++ IOTV/src/MK/FreerRTOS/id_5/ESP32/iotv/main/service/bme280/bme280.c
@@ -33,9 +33,9 @@ void BME280_service_task(void *pvParameters)
 		thp = BME280_readValues();
 		if (!thp.err)
 		{
-			*(double *)iot->readChannel[CH_TEMP].data = thp.temperature;
-			*(double *)iot->readChannel[CH_PRES].data = thp.pressure;
-			*(double *)iot->readChannel[CH_HUM].data = thp.humidity;
+			*(float *)iot->readChannel[CH_TEMP].data = thp.temperature;
+			*(float *)iot->readChannel[CH_PRES].data = thp.pressure;
+			*(float *)iot->readChannel[CH_HUM].data = thp.humidity;
 		}
 
 		vTaskDelay(5000 / portTICK_PERIOD_MS);
